Moved type to util folder and export
