+++ src/util/types.tsx
@@ -1,0 +1,9 @@
+export type ScreeningType = {
+    id: string;
+    date: string;
+    title: string;
+    location: string;
+    spokenLang: string;
+    subLang: string;
+    poster: string;
+  };
